chore: rename queries og-image -> og
